Initialized version control. Have created package gen which will attempt to find classes with the @Expose attribute, and generate the appropriate test code to allow the testing of private members of a class.
Very little testing is currently done, and limitations include:
Cannot accept primitives as arguments. This is because they are
a special case in the java byte-code (method.invoke() returns an
object, and the return type is wrong if they just get casted to
their boxed versions).

More to come. Example use cases (i.e. tests) are next on the list.